fixed some warnings and the FMOD absolute directory problem

+++ Engine/Code/Engine/Audio/AudioCueDefinition.cpp
@@ -26,6 +26,6 @@ AudioCueDefinition::AudioCueDefinition( const tinyxml2::XMLElement& xml ) {
 
 
 SoundID AudioCueDefinition::GetRandomSoundFromCue() const {
-	int soundIndex = GetRandomIntLessThan(m_sounds.size());
+	int soundIndex = GetRandomIntLessThan((int) m_sounds.size());
 	return m_sounds[soundIndex];
 }
